scripts/requirements.txt
requests~=2.22.0
PyYAML~=5.3.1
